tests: drop the one %ld in each interpreter's -v diagnostic; stop forcing float/long printf support on them
pint.c, adaint.c, fint.c, and forint.c each had exactly one printf-family
call using %ld, inside a -v usage-summary diagnostic that isn't even
exercised by the regression harness's normal args. slen (and its
per-app equivalents) is a long only because MAXSRC exceeds int range,
but the value is always bounds-checked against MAXSRC before use
(50000/9000/8000/7000, all well under 65535), so printing it via %u
with an (unsigned) cast is exact and avoids needing long-format printf
support at all. Verified each program's -v output is byte-identical to
before (e.g. "640 / 50000"), and confirmed via grep that every
printf-family call site in all four files uses a literal format string
directly (no wrapper functions, no va_list, nothing non-literal) - so
dcc's per-call auto-detection was already sufficient and this change is
safe.

Also added "dcc_floatio": false, "dcc_longio": false overrides for all
four in _test_overrides.json, matching the precedent already set by
cobint (which never needed them either). Get-DccFloatio/Get-DccLongio
default to true for every app unless overridden, so without this the
regression harness was forcing the entire float- and long-format printf
subsystem to stay linked into these four binaries regardless of the
source fix, masking the improvement in perf_baselines.csv. With the
overrides in place the real savings show up: peep/nopeep sizes drop by
roughly 3200-3328 bytes per app (pint, adaint, fint, forint all show
the same magnitude, since none of them use %f either and removing the
forced dcc-floatio=true eliminates that RTL machinery too, not just the
long-format piece the source edit alone accounts for).

Full regression suite (335 apps, both peep/nopeep, including each
app's ttt/sieve extra_scenarios) passes with zero failures.

+++ tests/fint.c
@@ -785,7 +785,7 @@ static void print_stats(void)
 {
     if (!opt_verbose) return;
     fprintf(stderr, "\nFINT usage summary\n");
-    fprintf(stderr, "  Source bytes:          %ld / %ld\n", slen, MAXSRC);
+    fprintf(stderr, "  Source bytes:          %u / %u\n", (unsigned)slen, (unsigned)MAXSRC);
     fprintf(stderr, "  Bytecode instructions: %d / %d\n", cp, MAXCODE);
     fprintf(stderr, "  Words:                 %d / %d\n", nw, MAXWORDS);
     fprintf(stderr, "  Strings:               %d / %d\n", ns, MAXSTR);
